+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.2.0 under development
 
-- New #90: Add method `itemsAsValues()` to widgets `RadioList` and `CheckboxList` that set items with labels equal
+- New #90: Add method `itemsFromValues()` to widgets `RadioList` and `CheckboxList` that set items with labels equal
   to values (vjik)
 - New #92: A third optional argument `$attributes` containing tag attributes in terms of name-value pairs has been
   added to methods `Html::textInput()`, `Html::hiddenInput()`, `Html::passwordInput()`, `Html::fileInput()`,